Change some files

+++ demo.cpp
@@ -5,7 +5,7 @@ int main(){
     int n,a[105],dp[105];
     cin>>n;
     for (int i=1;i<=n;i++) cin>>a[i];
-    ll ans=0;
+    int ans=0;
     for (int i=1;i<=n;i++){
         dp[i]=1;
         for (int j=1;j<i;j++){
